Remove 64bit warnings from 753

+++ far/panel.cpp
@@ -2089,7 +2089,7 @@ int Panel::SetPluginCommand(int Command,int Param1,LONG_PTR Param2)
 				}
 				if(Param1&&Param2)
 					xwcsncpy((wchar_t*)Param2,strInfoCurDir,Param1-1);
-				Result=strInfoCurDir.GetLength()+1;
+        Result=(int)strInfoCurDir.GetLength()+1;
 			}
 			break;
 
@@ -2103,13 +2103,13 @@ int Panel::SetPluginCommand(int Command,int Param1,LONG_PTR Param2)
 				{
 					if(Param1&&Param2)
 						xwcsncpy((wchar_t*)Param2,strColumnTypes,Param1-1);
-					Result=strColumnTypes.GetLength()+1;
+          Result=(int)strColumnTypes.GetLength()+1;
 				}
 				else
 				{
 					if(Param1&&Param2)
 						xwcsncpy((wchar_t*)Param2,strColumnWidths,Param1-1);
-					Result=strColumnWidths.GetLength()+1;
+          Result=(int)strColumnWidths.GetLength()+1;
 				}
 			}
 			break;
